refactored wallet screen
