+++ lib/src/app.dart
@@ -1,10 +1,12 @@
-import 'package:ea_kazi/src/features/messages_ui_screen.dart/messages_ui_screen.dart';
+import 'package:ea_kazi/src/features/features_exports.dart';
+import 'package:ea_kazi/src/features/wallet/presentation/withdraw_screen.dart';
 import 'package:ea_kazi/src/routes/router.gr.dart';
 import 'package:ea_kazi/src/utils/utils_exports.dart';
 import 'package:flutter/material.dart';
 
 
 //TODO: REMEBER TO REMOVE VIDEOS FOLDER FROM ASSETS AND PUBSPEC
+//TODO: IMPLEMENT API DATA IN REUSABLE WIDGET 
 
 class EaKazi extends StatelessWidget {
   const EaKazi({Key? key}) : super(key: key);
@@ -16,7 +18,7 @@ class EaKazi extends StatelessWidget {
     return MaterialApp(
       theme: theme,
       debugShowCheckedModeBanner: false,
-      home: MessagesUIScreen(),
+      home: WithdrawScreen(),
       // routerDelegate: appRouter.delegate(),
       // routeInformationParser: appRouter.defaultRouteParser(),
     );
