+++ lib/src/common/common_exports.dart
@@ -4,3 +4,4 @@ export 'search_field.dart';
 export 'reusable_card_widget.dart';
 export 'ea_kazi_app_bar.dart';
 export 'custom_divider.dart';
+export 'reusable_widget.dart';
