+++ lib/src/features/features_exports.dart
@@ -1,1 +1,2 @@
 export 'authentication/authentication_exports.dart';
+export 'wallet/wallet_exports.dart';
