+++ lib/src/features/wallet/wallet_exports.dart
@@ -1,0 +1,3 @@
+export 'presentation/wallet_screen.dart';
+export 'presentation/deposit_screen.dart';
+export 'presentation/withdraw_screen.dart';
